+++ apps/northware-cockpit/package.json
@@ -11,6 +11,7 @@
   "dependencies": {
     "@neondatabase/serverless": "^0.9.5",
     "@northware/ui": "workspace:*",
+    "@northware/auth": "workspace:*",
     "bcryptjs": "^2.4.3",
     "drizzle-orm": "^0.33.0",
     "next": "14.2.11",
