+++ apps/northware-cockpit/src/app/login/page.jsx
@@ -1,6 +1,6 @@
 import Link from "next/link";
 import { Form } from "../form";
-import { signIn } from "../auth";
+import { signIn } from "@northware/auth/auth";
 import { SubmitButton } from "../submit-button";
 
 export default function Login() {
