+++ apps/northware-cockpit/src/middleware.js
@@ -1,5 +1,5 @@
 import NextAuth from "next-auth";
-import { authConfig } from "./app/auth.config";
+import { authConfig } from "@northware/auth/authConfig";
 
 export default NextAuth(authConfig).auth;
