+++ packages/auth/README.md
@@ -1,0 +1,1 @@
+# `@northware/auth`
