+++ packages/auth/package.json
@@ -1,0 +1,15 @@
+{
+  "name": "@northware/auth",
+  "version": "0.0.0",
+  "private": true,
+  "dependencies": {
+    "@northware/database": "workspace:*",
+    "bcrypt": "^5.1.1",
+    "next-auth": "5.0.0-beta.20"
+  },
+  "exports": {
+    "./auth": "./src/auth.js",
+    "./user": "./src/user-connect.js",
+    "./authConfig": "./src/auth.config.js"
+  }
+}
