+++ packages/auth/src/auth.config.js
