+++ packages/auth/src/auth.js
@@ -1,4 +1,4 @@
-import { getUser } from "@/user-connect";
+import { getUser } from "./user-connect";
 import NextAuth from "next-auth";
 import Credentials from "next-auth/providers/credentials";
 import { authConfig } from "./auth.config";
